Fail if jni dylib is stale
