+++ jni/test/global_env_test.dart
@@ -25,6 +25,7 @@ void main() {
   // You have to manually pass the path to the `dartjni` dynamic library.
 
   if (!Platform.isAndroid) {
+    checkDylibIsUpToDate();
     try {
       Jni.spawn(dylibDir: "build/jni_libs", jvmOptions: ["-Xmx128m"]);
     } on JvmExistsException catch (_) {
