+++ jni/test/type_test.dart
@@ -205,6 +205,7 @@ class $FType extends JObjType<F> {
 
 void main() {
   if (!Platform.isAndroid) {
+    checkDylibIsUpToDate();
     try {
       Jni.spawn(dylibDir: "build/jni_libs", jvmOptions: ["-Xmx128m"]);
     } on JvmExistsException catch (_) {
